+++ src/config_old.rs
@@ -2,6 +2,29 @@ use serde_derive::{Deserialize, Serialize};
 
 use crate::{config::GfxConfig, gfx_vendors::GfxMode};
 
+#[derive(Debug, PartialEq, Copy, Clone, Deserialize, Serialize)]
+pub enum GfxMode300 {
+    Hybrid,
+    Nvidia,
+    Integrated,
+    Compute,
+    Vfio,
+    Egpu,
+}
+
+impl From<GfxMode300> for GfxMode {
+    fn from(m: GfxMode300) -> Self {
+        match m {
+            GfxMode300::Hybrid => GfxMode::Hybrid,
+            GfxMode300::Nvidia => GfxMode::Dedicated,
+            GfxMode300::Integrated => GfxMode::Integrated,
+            GfxMode300::Compute => GfxMode::Compute,
+            GfxMode300::Vfio => GfxMode::Vfio,
+            GfxMode300::Egpu => GfxMode::Egpu,
+        }
+    }
+}
+
 #[derive(Deserialize, Serialize)]
 pub struct GfxConfig300 {
     #[serde(skip)]
@@ -21,7 +44,7 @@ impl From<GfxConfig300> for GfxConfig {
     fn from(old: GfxConfig300) -> Self {
         GfxConfig {
             config_path: old.config_path,
-            mode: old.gfx_mode,
+            mode: old.gfx_mode.into(),
             tmp_mode: old.gfx_tmp_mode,
             vfio_enable: old.gfx_vfio_enable,
             vfio_save: false,